+++ backend/src/app/sdf-controller.ts
@@ -20,18 +20,23 @@ import { RuleEvaluator } from './rule-engine';
 import DV360Facade from './dv360-facade';
 import SdfService, { GenerateSdfOptions } from './sdf-service';
 import { FeedData } from '../types/types';
+import { getCurrentDateTimestamp } from './utils';
+import { Logger } from '../types/logger';
 
 export default class SdfController {
-  constructor(private configService: ConfigService,
+  constructor(
+    private logger: Logger,
+    private configService: ConfigService,
     private ruleEvaluator: RuleEvaluator,
     private feedService: FeedService,
     private dvFacade: DV360Facade) {
+    if (!logger) throw new Error('[SdfController] ArgumentException: Required argument logger is missing');
   }
   config?: Config;
 
   async generateSdf(configId: string, options: GenerateSdfOptions): Promise<string> {
     options = options || {};
-    console.log(`[SdfController] Generating SDF for configuration ${configId} with options:\n ${JSON.stringify(options, null, 2)}`);
+    this.logger.info(`[SdfController] Generating SDF for configuration ${configId} with options:\n ${JSON.stringify(options, null, 2)}`);
     if (!options.update) {
       if (!options.startDate)
         throw new Error(`[SdfController] Campaign start date should be specified`);
@@ -44,25 +49,27 @@ export default class SdfController {
     try {
       config = await this.configService.loadConfiguration(configId);
     } catch (e) {
-      console.error(`[SdfController] Fetching configuration failed: ${e.message}`);
+      this.logger.error(`[SdfController] Fetching configuration failed: ${e.message}`);
+      e.logged = true;
       throw e;
     }
     this.config = config;
 
     // #2 Validate configuration
-    let errors = this.configService.validateGeneratingConfiguration(config, !!options.update);
+    let errors = ConfigService.validateGeneratingConfiguration(config, !!options.update);
     if (errors && errors.length)
       throw new Error(`[SdfController] There are errors in configuration that prevents from generating SDFs:\n` +
         errors.map(e => e.message).join(',\n')
       );
 
     // #3 Load data from feed(s)
-    let sdfSvc = new SdfService(config, this.ruleEvaluator, this.dvFacade);
+    let sdfSvc = new SdfService(config, this.logger, this.ruleEvaluator, this.dvFacade);
     let feedData: FeedData;
     try {
       feedData = await this.feedService.loadAll(config.feedInfo!);
     } catch (e) {
-      console.error(`[SdfController] Fetching combined feed failed: ${e.message}`);
+      this.logger.error(`[SdfController] Fetching combined feed failed: ${e.message}`);
+      e.logged = true;
       throw e;
     }
     if (feedData.rowCount === 0) {
@@ -70,7 +77,7 @@ export default class SdfController {
     }
 
     // #4: validate columns for first row
-    errors = this.configService.validateGeneratingRuntimeConfiguration(config, feedData);
+    errors = ConfigService.validateGeneratingRuntimeConfiguration(config, feedData);
     if (errors && errors.length)
     throw new Error(`[SdfController] There are errors in configuration that prevents from generating SDFs:\n` +
       errors.map(e => e.message).join(',\n')
@@ -79,12 +86,13 @@ export default class SdfController {
     // #5 Generate SDF
     let filepath: string;
     if (!options.fileName) {
-      options.fileName = "sdf-" + new Date().toISOString().replace(/\-/g, "").replace(/\:/g, "").replace(".", "") + ".zip";
+      options.fileName = "sdf-" + getCurrentDateTimestamp() + ".zip";
     }
     try {
       filepath = await sdfSvc.generateSdf(feedData, options);
     } catch (e) {
-      console.error(`[SdfController] Generating SDF failed: ${e.message}`);
+      this.logger.error(`[SdfController] Generating SDF failed: ${e.message}`);
+      e.logged = true;
       throw e;
     }
     return filepath;