+++ backend/src/argv.ts
@@ -1,0 +1,1 @@
+export = require('yargs/yargs')(process.argv.slice(2)).argv;